remove health check endpoints through Graphql seperate example trip schema and resolver from the default ones
