+++ templates/package.json
@@ -20,6 +20,7 @@
     "apollo-datasource-rest": "^0.9.7",
     "apollo-server-express": "^2.19.2",
     "graphql": "^15.5.0",
+    "graphql-combine": "^1.0.1",
     "isemail": "^3.1.3",
     <% end %>
     "aws-cloudfront-sign": "^2.2.0",
