+++ templates/src/graphql/default.graphql
@@ -1,0 +1,43 @@
+"User object for the authentication"
+type User {
+  """
+  The user id that Identity Provider assigned.
+  """
+  id: ID!
+  email: String!
+}
+"Presigned url object"
+type PresignedUrl {
+  """
+  A presigned url generated by a cloud storage points to a specific file 
+  that has been stored or will be stored on a cloud storage
+  """
+  url: String
+  """
+  The http request method such as GET, PUT, POST
+  """
+  method: String 
+}
+"A presigned urls collection that is able to return all presigned urls to a user at the same time. "
+type PresignedUrls {
+  """
+  The presigned url object for file uploading
+  """
+  upload: PresignedUrl
+  """
+  The presigned url object for file downloading
+  """
+  download: PresignedUrl
+}
+
+"Type Query defines all query / read methods that clients can call. These methods must be implemented in resolver as well."
+type Query {
+  """
+  Get all presigned urls for a specific file that has been stored or will be stored on a cloud storage
+  """
+  presignedUrls(key: String!): PresignedUrls
+  """
+  Get user object from the decoded JWT
+  """
+  userInfo: User,
+}
